+++ src/OpenTelemetry.Instrumentation.AspNet/AspNetInstrumentation.cs
@@ -1,6 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+using System.Diagnostics;
 using System.Diagnostics.Metrics;
 using System.Reflection;
 using OpenTelemetry.Instrumentation.AspNet.Implementation;
@@ -18,7 +19,9 @@ internal sealed class AspNetInstrumentation : IDisposable
     public static readonly Assembly Assembly = typeof(HttpInListener).Assembly;
     public static readonly AssemblyName AssemblyName = Assembly.GetName();
     public static readonly string MeterName = AssemblyName.Name!;
+    public static readonly string ActivitySourceName = AssemblyName.Name;
     public static readonly Meter Meter = new(MeterName, Assembly.GetPackageVersion());
+    public static readonly ActivitySource ActivitySource = new(ActivitySourceName, Assembly.GetPackageVersion());
     public static readonly Histogram<double> HttpServerDuration = Meter.CreateHistogram(
         "http.server.request.duration",
         unit: "s",