+++ src/OpenTelemetry.Instrumentation.AspNet.TelemetryHttpModule/CHANGELOG.md
@@ -122,7 +122,7 @@ Released 2023-Feb-27
 
 Released 2022-Nov-28
 
-* Restore Activity.Current before all IIS Lifecycle events
+* Restore Activity.Current before all IIS Lifecycle events.
   ([#761](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/761))
 
 ## 1.0.0-rc9.6