+++ src/OpenTelemetry.Instrumentation.AspNet.TelemetryHttpModule/CHANGELOG.md
@@ -4,8 +4,8 @@
 
 * **Breaking Change**: This module is no longer responsible for creating activities.
   The contract of `TelemetryHttpModuleOptions.OnRequestStartedCallback` was changed
-  to `Func<HttpContextBase, ActivityContext, Activity?>?`. It means that the consumer
-  is responsible for providing function returning `Activity`.
+  to `Func<HttpContextBase, ActivityContext, Activity?>?`. The consumer is now
+  responsible for providing function returning `Activity`.
   ([#3151](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3151))
 
 ## 1.12.0-beta.2